Remove need to inject in MuleEvent as argument for seven api methods
Use RequestContext.getEvent() to supply the MuleEvent object for the
seven api methods in RedisModule class, instead of needing it as the
last method argument.

Although RequestContext is deprecated, it is still used in the RedisModule
class in the private method getActualDefaultPartitionName(), so we're not
introducing any new use of that class.

We don't know any other reasonable way to supply the MuleEvent object in
the flow xml, once we remove the @Inject annotation for those api methods.

The @Inject annotation causes our project maven build's integration tests
to all fail with:
 BeanCreationException: Could not autowire method: ...RedisModule.set(...)
The deprecation comments suggest implementing MessageProcessor or Callable
to get access to the MuleEvent, but that is not applicable in our case
because we're not currently using redis connector in an xml flow. For more
'blow-by-blow' background, see the (long) discussion thread in
 https://github.com/mulesoft/redis-connector/issues/3

Remove the interim previously needed fix to the
 mule-module-redis.xml.sample
for 'mvn clean package' to not fail.